+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v64.0.1...main)
+[Full changelog](https://github.com/mozilla/glean/compare/v64.1.0...main)
+
+# v64.1.0 (2025-04-07)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v64.0.1...v64.1.0)
+
+* General
+  * FEATURE: New client-wide attribution and distribution fields ([bug 1955428](https://bugzilla.mozilla.org/show_bug.cgi?id=1955428))
+  * Update to `glean_parser` v17.1.0
+* Kotlin
+   * Updated Android Gradle Plugin to 8.9.1 ([#3098](https://github.com/mozilla/glean/pull/3098))
+   * Updated Kotlin to version 2.1.20 ([#3098](https://github.com/mozilla/glean/pull/3098))
+   * Dispatch ping API on the task queue ([#3101](https://github.com/mozilla/glean/pull/3101))
 
 # v64.0.1 (2025-04-01)